fix: use i64 for SEQ_IN_INDEX to handle BIGINT from MySQL/MariaDB
Some MySQL/MariaDB versions return SEQ_IN_INDEX as BIGINT rather than
INT UNSIGNED, causing a type mismatch error at decode time. Change the
Rust type from u32 to i64 across fetcher, schema, and sql_generator.

+++ src/fetcher.rs
@@ -207,7 +207,7 @@ async fn fetch_indexes(
     for row in rows {
         let table_name: String = row.try_get("TABLE_NAME")?;
         let index_name: String = row.try_get("INDEX_NAME")?;
-        let seq: u32 = row.try_get("SEQ_IN_INDEX")?;
+        let seq: i64 = row.try_get("SEQ_IN_INDEX")?;
         let non_unique: i32 = row.try_get("NON_UNIQUE")?;
         let index_type: String = row.try_get("INDEX_TYPE")?;
         let sub_part: Option<i64> = row.try_get("SUB_PART")?;
